Pan Tilt RF Control Laser
This is a program I spent all night on.  It is hilarious how I haven't slept in the past few days, but I have been energetic and productive as ever.

Basically, hold the RF dongle so that "B" is up and "C" is down.  
Use the dongle like a video game controller.

Press the up/down at the same time to turn on the Laser.
Or Press the Left & Right Buttons simultaneously to turn off the laser.